+++ src/app/dim-ui/ActivityTracker.tsx
@@ -3,8 +3,8 @@ import * as React from 'react';
 import { $rootScope } from 'ngimport';
 import { loadingTracker } from '../ngimport-more';
 
-const ONE_MINUTE = 60 * 1000;
-const FIVE_MINUTES = 5 * 60 * 1000;
+const MIN_REFRESH_INTERVAL = 10 * 1000;
+const AUTO_REFRESH_INTERVAL = 30 * 1000;
 const ONE_HOUR = 60 * 60 * 1000;
 
 /**
@@ -22,7 +22,7 @@ export class ActivityTracker extends React.Component {
     // This event should *NOT* be listened to by services!
     // TODO: replace this with an observable?
     $rootScope.$broadcast('dim-refresh');
-  }, ONE_MINUTE, { trailing: false });
+  }, MIN_REFRESH_INTERVAL, { trailing: false });
 
   componentDidMount() {
     this.track();
@@ -58,9 +58,8 @@ export class ActivityTracker extends React.Component {
     return (Date.now() - this.lastActivityTimestamp) <= timespan;
   }
 
-  // Refresh every 5 minutes automatically
   private startTimer() {
-    this.refreshAccountDataInterval = window.setTimeout(this.refreshAccountData, FIVE_MINUTES);
+    this.refreshAccountDataInterval = window.setTimeout(this.refreshAccountData, AUTO_REFRESH_INTERVAL);
   }
 
   private clearTimer() {