+++ src/app/shell/Link.tsx
@@ -9,6 +9,7 @@ interface Props {
   state: string;
   text?: string;
   children?: React.ReactChild;
+  showWhatsNew?: boolean;
 }
 
 interface State {
@@ -41,14 +42,15 @@ export default class Link extends React.Component<Props, State> {
   }
 
   render() {
-    const { state, account, children, text } = this.props;
+    const { state, account, children, text, showWhatsNew } = this.props;
 
     // This should be a really simple component, but because of https://github.com/ui-router/react/issues/204
     // it can't handle lazy states, and we need to use "key" to nuke the whole component tree on updates.
+
     return (
       <UISrefActive key={this.state.generation} class='active'>
         <UISref to={state} params={account}>
-          <a className='link'>{children}{text && t(text)}</a>
+          <a className='link'>{showWhatsNew && <span className='badge-new'/>}{children}{text && t(text)}</a>
         </UISref>
       </UISrefActive>
     );